created
Note: buttonlib isn't actually used here, it is the result of taking
some of the things I created for this project and putting them in to a
more general-case button framework. Eventually I'll get around to making
map_view use buttonlib instead of its own version.

Things todo:
Add the ability to load tilesets
add the ability to set the width/height of new maps
add the ability to set the width/height of current maps